moved cities and animal list to context
